Endpoint 4 working

+++ applications/order/src/main/java/com/example/OrderLine/OrderLine.java
@@ -24,12 +24,12 @@ public class OrderLine {
     private Double quantity;
     private Double price;
     private Double totalPrice;
-    private Double shipmentId;
+    private Integer shipmentId;
     @ManyToOne
     @JsonBackReference
     private Order order;
 
-    public OrderLine(Order order, Integer productId, Double quantity, Double price, Double shipmentId){
+    public OrderLine(Order order, Integer productId, Double quantity, Double price, Integer shipmentId){
         this.order = order;
         this.productId = productId;
         this.quantity = quantity;
@@ -37,7 +37,7 @@ public class OrderLine {
         this.totalPrice = quantity * price;
         this.shipmentId = shipmentId;
     }
-    public OrderLine(Integer productId, Double quantity, Double price, Double shipmentId){
+    public OrderLine(Integer productId, Double quantity, Double price, Integer shipmentId){
         this.productId = productId;
         this.quantity = quantity;
         this.price = price;
